+++ CheckEat-User/ViewModels/Home/StoreMapViewModel.swift
