+++ CheckEat-User/Views/Home/Favorites/FavoriteStoreScreenView.swift
